backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.3
SQLAlchemy==2.0.35
python-dotenv==1.0.1
gunicorn==23.0.0
logging==0.4.9.6
